ACT-438: Clean up admin models in form library app
- Move all admin models to admin.py
- Use decorators to register admin models
