+++ formlibrary/admin.py
@@ -3,10 +3,27 @@
 from django.contrib import admin
 
 from .models import (
-    TrainingAttendance, TrainingAttendanceAdmin, Distribution,
-    DistributionAdmin, Beneficiary, BeneficiaryAdmin
+    TrainingAttendance, Distribution, Beneficiary
 )
 
-admin.site.register(TrainingAttendance, TrainingAttendanceAdmin)
-admin.site.register(Distribution, DistributionAdmin)
-admin.site.register(Beneficiary, BeneficiaryAdmin)
+
+@admin.register(Beneficiary)
+class BeneficiaryAdmin(admin.ModelAdmin):
+    list_display = ('site', 'beneficiary_name',)
+    display = 'Beneficiary'
+    list_filter = ('site', 'beneficiary_name')
+
+
+@admin.register(Distribution)
+class DistributionAdmin(admin.ModelAdmin):
+    list_display = ('distribution_name', 'program',
+                    'initiation', 'create_date', 'edit_date')
+    display = 'Program Dashboard'
+
+
+@admin.register(TrainingAttendance)
+class TrainingAttendanceAdmin(admin.ModelAdmin):
+    list_display = ('training_name', 'program',
+                    'project_agreement', 'create_date', 'edit_date')
+    display = 'Training Attendance'
+    list_filter = ('program__country', 'program')
